requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
aiosqlite>=0.20.0
pytest>=8.0.0
httpx>=0.28.0
pytest-asyncio>=0.24.0
